replacing redux help-block logic
