+++ src/redux/user/operation.js
@@ -13,15 +13,5 @@ export const updateUser = createAsyncThunk(
     }
 )
 
-export const requestHelp = createAsyncThunk(
-  'users/help/request',
-  async (credentials, thunkAPI) => {
-    try {
-      const res = await API.support(credentials);
-      return res.data;
-    } catch (err) {
-      return thunkAPI.rejectWithValue(err.message);
-    }
-  }
-);
+
