+++ src/redux/user/slice.js
@@ -1,6 +1,5 @@
 const { createSlice } = require('@reduxjs/toolkit');
 const { updateUser } = require('./operation');
-import { requestHelp } from './operation.js';
 
 const handlePending = state => {
   state.isLoading = true;
@@ -40,30 +39,3 @@ const userSlice = createSlice({
 
 export const userReducer = userSlice.reducer;
 
-export const helpSlice = createSlice({
-  name: 'help',
-  initialState: {
-    email: '',
-    comment: '',
-    isLoading: false,
-    error: null,
-  },
-  extraReducers: builder => {
-    builder
-      .addCase(requestHelp.pending, state => {
-        state.isLoading = true;
-      })
-      .addCase(requestHelp.fulfilled, (state, { payload }) => {
-        state.email = payload;
-        state.comment = payload;
-        state.isLoading = false;
-        state.error = null;
-      })
-      .addCase(requestHelp.rejected, (state, action) => {
-        state.isLoading = false;
-        state.error = action.payload;
-      });
-  },
-});
-
-export const requestHelpReducer = helpSlice.reducer;
